fixes #1083 and creates totalPages prop to iterate less and avoid massive arrays for large pagination
