+++ packages/ui/src/components/pagination/components/pagination-presentational/pagination-presentational.types.ts
@@ -1,0 +1,46 @@
+import { HTMLAttributes } from 'react';
+
+import { type PaginationItemProps } from '../../components/index.js';
+import { PaginationBase } from '../../pagination.types.js';
+
+export type PageToRender =
+  | {
+      href: string;
+      page: number;
+      text: React.ReactNode;
+    }
+  | {
+      page: number;
+      text: React.ReactNode;
+    }
+  | null;
+
+export type PaginationPresentationalProps = PaginationBase & {
+  /**
+   * on page change
+   */
+  onChange?: (page: number) => unknown;
+  /**
+   * Pages to render where the null value is rendered as [...]
+   */
+  pagesToRender: PageToRender[];
+  /**
+   * Props to back button
+   */
+
+  paginationBackProps?: PaginationItemProps;
+  /**
+   * Props to back button
+   */
+  paginationNextProps?: PaginationItemProps;
+  /**
+   * siblingCount
+   * @default 1
+   */
+  siblingCount?: number;
+  /**
+   * Tag to render
+   * @default nav
+   */
+  tag?: keyof JSX.IntrinsicElements;
+} & Omit<HTMLAttributes<Element>, 'onChange'>;
